pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.121.3",
    "psycopg2-binary>=2.9.11",
    "pydantic>=2.12.4",
    "uvicorn[standard]>=0.38.0",
]
